Created initial repository and first demo.
This repository holds demo applications to teach students how
to exploit modern computational resources and tools to solve
problems in Economics.  This commit creates the first demo
application which is demonstrates how to show that an estimator
is asmptotically unbiased by exploiting the fact that estimating
multiple different datasets is embarrassingly parallel.